moved schema into separate files
