+++ __test__/lib.test.js
@@ -1,4 +1,5 @@
-import { schema, resolvers } from '../src/server/api/schema'
+import { resolvers } from '../src/server/api/schema'
+import { schema } from '../src/api/schema'
 import twilio from '../src/server/api/lib/twilio'
 import { makeExecutableSchema } from 'graphql-tools'
