+++ src/api/assignment.js
@@ -1,0 +1,15 @@
+export const schema = `
+  input AssignmentsFilter {
+    texterId: Int
+  }
+  type Assignment {
+    id: ID
+    texter: User
+    campaign: Campaign
+    contacts(contactsFilter: ContactsFilter): [CampaignContact]
+    contactsCount(contactsFilter: ContactsFilter): Int
+    userCannedResponses: [CannedResponse]
+    campaignCannedResponses: [CannedResponse]
+    maxContacts: Int
+  }
+`
