+++ src/api/campaign-contact.js
@@ -1,0 +1,41 @@
+export const schema = `
+  input ContactsFilter {
+    messageStatus: String
+    isOptedOut: Boolean
+    validTimezone: Boolean
+    includePastDue: Boolean
+  }
+
+  type Timezone {
+    offset: Int
+    hasDST: Boolean
+  }
+
+  type Location {
+    timezone: Timezone
+    city: String
+    state: String
+  }
+
+  type CampaignContact {
+    id: ID
+    firstName: String
+    lastName: String
+    cell: Phone
+    zip: String
+    external_id: String
+    customFields: JSON
+    messages: [Message]
+    location: Location
+    optOut: OptOut
+    campaign: Campaign
+    questionResponseValues: [AnswerOption]
+    questionResponses: [AnswerOption]
+    interactionSteps: [InteractionStep]
+    currentInteractionStepScript: String
+    currentInteractionStepId: String
+    messageStatus: String
+    assignmentId: String
+  }
+`
+
