+++ src/api/conversations.js
@@ -1,0 +1,18 @@
+export const schema = `
+  input ConversationFilter {
+    assignmentsFilter: AssignmentsFilter
+    campaignsFilter: CampaignsFilter
+    contactsFilter: ContactsFilter
+  }
+
+  type Conversation {
+    texter: User!
+    contact: CampaignContact!
+    campaign: Campaign!
+  }
+  
+  type PaginatedConversations {
+    conversations: [Conversation]!
+    pageInfo: PageInfo
+  }
+`
