+++ src/api/invite.js
@@ -1,0 +1,8 @@
+export const schema = `
+  type Invite {
+    id: ID
+    isValid: Boolean
+    hash: String
+  }
+`
+
