+++ src/api/opt-out.js
@@ -1,0 +1,9 @@
+export const schema = `
+  type OptOut {
+    id: ID
+    cell: String
+    assignment: Assignment
+    createdAt: Date
+  }
+`
+
