+++ src/api/question-response.js
@@ -1,0 +1,8 @@
+export const schema = `
+  type QuestionResponse {
+    id: String
+    value: String
+    question: Question
+  }
+`
+
