+++ src/server/api/assignment.js
@@ -2,21 +2,6 @@ import { mapFieldsToModel } from './lib/utils'
 import { Assignment, r } from '../models'
 import { getOffsets, defaultTimezoneIsBetweenTextingHours } from '../../lib'
 
-export const schema = `
-  input AssignmentsFilter {
-    texterId: Int
-  }
-  type Assignment {
-    id: ID
-    texter: User
-    campaign: Campaign
-    contacts(contactsFilter: ContactsFilter): [CampaignContact]
-    contactsCount(contactsFilter: ContactsFilter): Int
-    userCannedResponses: [CannedResponse]
-    campaignCannedResponses: [CannedResponse]
-    maxContacts: Int
-  }
-`
 export function addWhereClauseForContactsFilterMessageStatusIrrespectiveOfPastDue(
   queryParameter,
   contactsFilter
