+++ src/server/api/canned-response.js
@@ -1,25 +1,6 @@
 import { mapFieldsToModel } from './lib/utils'
 import { CannedResponse } from '../models'
 
-export const schema = `
-  input CannedResponseInput {
-    id: String
-    title: String
-    text: String
-    campaignId: String
-    userId: String
-  }
-
-  type CannedResponse {
-    id: ID
-    title: String
-    text: String
-    isUserCreated: Boolean
-    campaign: Campaign
-    user: User
-  }
-`
-
 export const resolvers = {
   CannedResponse: {
     ...mapFieldsToModel([
