+++ src/server/api/interaction-step.js
@@ -1,20 +1,6 @@
 import { mapFieldsToModel } from './lib/utils'
 import { InteractionStep, r } from '../models'
 
-export const schema = `
-  type InteractionStep {
-    id: ID!
-    question: Question
-    questionText: String
-    script: String
-    answerOption: String
-    parentInteractionId: String
-    isDeleted: Boolean
-    answerActions: String
-    questionResponse(campaignContactId: String): QuestionResponse
-  }
-`
-
 export const resolvers = {
   InteractionStep: {
     ...mapFieldsToModel([
