+++ src/server/api/opt-out.js
@@ -1,15 +1,6 @@
 import { mapFieldsToModel } from './lib/utils'
 import { OptOut } from '../models'
 
-export const schema = `
-  type OptOut {
-    id: ID
-    cell: String
-    assignment: Assignment
-    createdAt: Date
-  }
-`
-
 export const resolvers = {
   OptOut: {
     ...mapFieldsToModel([
