+++ src/server/api/organization.js
@@ -4,25 +4,6 @@ import { accessRequired } from './errors'
 import { buildCampaignQuery } from './campaign'
 import { buildUserOrganizationQuery } from './user'
 
-export const schema = `
-  input PeopleFilter {
-    campaignsFilter: CampaignsFilter
-  }
-
-  type Organization {
-    id: ID
-    uuid: String
-    name: String
-    campaigns(campaignsFilter: CampaignsFilter): [Campaign]
-    people(role: String, campaignsFilter: CampaignsFilter): [User]
-    optOuts: [OptOut]
-    threeClickEnabled: Boolean
-    textingHoursEnforced: Boolean
-    textingHoursStart: Int
-    textingHoursEnd: Int
-  }
-`
-
 export const resolvers = {
   Organization: {
     ...mapFieldsToModel([
