+++ src/server/api/question-response.js
@@ -1,14 +1,6 @@
 import { mapFieldsToModel } from './lib/utils'
 import { QuestionResponse } from '../models'
 
-export const schema = `
-  type QuestionResponse {
-    id: String
-    value: String
-    question: Question
-  }
-`
-
 export const resolvers = {
   QuestionResponse: {
     ...mapFieldsToModel([
