+++ src/server/api/question.js
@@ -1,23 +1,5 @@
 import { r } from '../models'
 
-export const schema = `
-  type Question {
-    text: String
-    answerOptions: [AnswerOption]
-    interactionStep: InteractionStep
-  }
-
-  type AnswerOption {
-    interactionStepId: Int
-    value: String
-    action: String
-    nextInteractionStep: InteractionStep
-    responders: [CampaignContact]
-    responderCount: Int
-    question: Question
-  }
-`
-
 export const resolvers = {
   Question: {
     text: async (interactionStep) => interactionStep.question,
