+++ src/server/index.js
@@ -4,7 +4,8 @@ import express from 'express'
 import appRenderer from './middleware/app-renderer'
 import { graphqlExpress, graphiqlExpress } from 'apollo-server-express'
 import { makeExecutableSchema, addMockFunctionsToSchema } from 'graphql-tools'
-import { schema, resolvers } from './api/schema'
+import { resolvers } from './api/schema'
+import { schema } from '../api/schema'
 import { accessRequired } from './api/errors'
 import mocks from './api/mocks'
 import { createLoaders } from './models'
